refactor: move test files in same level as production code
