+++ src/Domain/Repository/CouponRepository.ts
@@ -1,4 +1,4 @@
-import Coupon from "../Entity/Coupon";
+import { Coupon } from "../Entity/Coupon";
 
 export interface CouponRepository {
   getByCode(code: string): Promise<Coupon>;
